# withings/openapi/generator/withings-swagger-v3.0.3-fixed.json
{
  "openapi": "3.0.3",
  "servers": [
    {
      "url": "https://wbsapi.withings.net"
    }
  ],
  "info": {
    "version": "2.0",
    "title": "Withings developer documentation",
    "x-logo": {
      "url": "/img/logo_withings.svg"
    },
...
  },
...
  "paths": {
...
    "/v2/oauth2?action=requesttoken": {
      "post": {
        "tags": [
          "oauth2"
        ],
        "description": "After you obtain the authorization code you can request an ```access_token```.\n\n**Important:** Must be called as POST call and parameters must be set in the body request as form-data parameters and not in the url.",
        "summary": "OAuth 2.0 - Get your access_token",
        "operationId": "oauth2-getaccesstoken",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your ```Client ID```."
          },
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DO NOT USE WITH FOLLOWING PARAMS: [client_secret]. A random token used to prevent replay attacks (Cf. [Signature v2 - Getnonce](#operation/signaturev2-getnonce))."
          },
          {
            "name": "signature",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DO NOT USE WITH FOLLOWING PARAMS: [client_secret]. Hash of params (Cf. [Signature hash protocol](/developer-guide/v3/get-access/sign-your-requests))."
          },
          {
            "name": "client_secret",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DO NOT USE WITH FOLLOWING PARAMS: [signature,nonce]. Your ```Client Secret```."
          },
          {
            "name": "grant_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Must take the constant string value ```authorization_code```."
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The Authorization Code you got from previous step."
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The URI you use in the first call."
          }
        ],
        "responses": {
          "200": {
            "description": " (Click to unfold)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "Response status. See <a href='#section/Response-status'>Status</a> section for details."
                    },
                    "body": {
                      "type": "object",
                      "description": "Response data.",
                      "properties": {
                        "userid": {
                          "type": "string",
                          "description": "The id of the user.",
                          "example": "363"
                        },
                        "access_token": {
                          "type": "string",
                          "description": "Your new Access Token.",
                          "example": "a075f8c14fb8df40b08ebc8508533dc332a6910a"
                        },
                        "refresh_token": {
                          "type": "string",
                          "description": "Your Refresh Token.",
                          "example": "f631236f02b991810feb774765b6ae8e6c6839ca"
                        },
                        "expires_in": {
                          "type": "integer",
                          "description": "Access token expiry delay in seconds.",
                          "example": 10800
                        },
                        "scope": {
                          "type": "string",
                          "description": "You can get only the scope that the user accepted with the Token you have. Scopes can be 'user.info' 'user.metrics' 'user.activity' 'user.sleepevents'  and must be separated by a coma.",
                          "example": "user.info,user.metrics"
                        },
                        "csrf_token": {
                          "type": "string",
                          "example": "PACnnxwHTaBQOzF7bQqwFUUotIuvtzSM"
                        },
                        "token_type": {
                          "type": "string",
                          "description": "HTTP Authorization Header format: Bearer",
                          "example": "Bearer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "PHP",
            "source": "<?php\n\n$ch = curl_init();\n\ncurl_setopt($ch, CURLOPT_URL, \"/v2/oauth2\");\n\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, TRUE);\n\ncurl_setopt($ch, CURLOPT_POSTFIELDS, http_build_query([ \n\t'action' => 'requesttoken',\n\t'grant_type' => 'authorization_code',\n\t'client_id' => '7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67',\n\t'client_secret' => 'd9286311451fc6ed71b372a075c58c5058be158f56a77865e43ab3783255424f',\n\t'code' => 'mtwsikawoqleuroqcluggflrqilrnqbgqvqeuhhh',\n\t'redirect_uri' => 'https://www.withings.com'\n]));\n\n$rsp = curl_exec($ch);\ncurl_close($ch);\n\nvar_dump($rsp);\n\n?>"
          },
          {
            "lang": "CURL",
            "source": "curl --data \"action=requesttoken&grant_type=authorization_code&client_id=7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67&client_secret=d9286311451fc6ed71b372a075c58c5058be158f56a77865e43ab3783255424f&code=mtwsikawoqleuroqcluggflrqilrnqbgqvqeuhhh&redirect_uri=https://www.withings.com\" 'https://wbsapi.withings.net/v2/oauth2'"
          }
        ]
      }
    },
    "/v2/oauth2?action=refreshaccesstoken": {
      "post": {
        "tags": [
          "oauth2"
        ],
        "description": "If user's ```access_token``` is expired, use this service with user's ```refresh_token``` to get a new ```access_token``` (and a new ```refresh_token```) and be able to make requests again.\n\n**Important:**\n\n - Must be called as POST call and parameters must be set in the body request as application/x-www-form-urlencoded parameters and not in the url.\n - This service also returns a new  ```refresh_token```. You have to overwrite your current  ```refresh_token``` with the new one.",
        "summary": "OAuth 2.0 - Refresh your access_token",
        "operationId": "oauth2-refreshaccesstoken",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your ```Client ID```."
          },
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DO NOT USE WITH FOLLOWING PARAMS: [client_secret]. A random token used to prevent replay attacks (Cf. [Signature v2 - Getnonce](#operation/signaturev2-getnonce))."
          },
          {
            "name": "signature",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DO NOT USE WITH FOLLOWING PARAMS: [client_secret]. Hash of params (Cf. [Signature hash protocol](/developer-guide/v3/get-access/sign-your-requests))."
          },
          {
            "name": "client_secret",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "DO NOT USE WITH FOLLOWING PARAMS: [signature,nonce]. Your ```Client Secret```."
          },
          {
            "name": "grant_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Must take the constant string value ```refresh_token```."
          },
          {
            "name": "refresh_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The current valid ```refresh_token```."
          }
        ],
        "responses": {
          "200": {
            "description": " (Click to unfold)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "Response status. See <a href='#section/Response-status'>Status</a> section for details."
                    },
                    "body": {
                      "type": "object",
                      "description": "Response data.",
                      "properties": {
                        "userid": {
                          "type": "string",
                          "description": "The id of the user.",
                          "example": "363"
                        },
                        "access_token": {
                          "type": "string",
                          "description": "Your new Access Token.",
                          "example": "a075f8c14fb8df40b08ebc8508533dc332a6910a"
                        },
                        "refresh_token": {
                          "type": "string",
                          "description": "Your Refresh Token.",
                          "example": "f631236f02b991810feb774765b6ae8e6c6839ca"
                        },
                        "expires_in": {
                          "type": "integer",
                          "description": "Access token expiry delay in seconds.",
                          "example": 10800
                        },
                        "scope": {
                          "type": "string",
                          "description": "You can get only the scope that the user accepted with the Token you have. Scopes can be 'user.info' 'user.metrics' 'user.activity' 'user.sleepevents'  and must be separated by a coma.",
                          "example": "user.info,user.metrics"
                        },
                        "csrf_token": {
                          "type": "string",
                          "example": "PACnnxwHTaBQOzF7bQqwFUUotIuvtzSM"
                        },
                        "token_type": {
                          "type": "string",
                          "description": "HTTP Authorization Header format: Bearer",
                          "example": "Bearer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "PHP",
            "source": "<?php\n\n$ch = curl_init();\n\ncurl_setopt($ch, CURLOPT_URL, \"/v2/oauth2\");\n\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, TRUE);\n\ncurl_setopt($ch, CURLOPT_POSTFIELDS, http_build_query([ \n\t'action' => 'requesttoken',\n\t'grant_type' => 'refresh_token',\n\t'client_id' => '7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67',\n\t'client_secret' => 'd9286311451fc6ed71b372a075c58c5058be158f56a77865e43ab3783255424f',\n\t'refresh_token' => '9697c3d06ccfd1ca302f5a527d345a9f99ea88a2'\n]));\n\n$rsp = curl_exec($ch);\ncurl_close($ch);\n\nvar_dump($rsp);\n\n?>"
          },
          {
            "lang": "CURL",
            "source": "curl --data \"action=requesttoken&grant_type=refresh_token&client_id=7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67&client_secret=d9286311451fc6ed71b372a075c58c5058be158f56a77865e43ab3783255424f&refresh_token=9697c3d06ccfd1ca302f5a527d345a9f99ea88a2\" 'https://wbsapi.withings.net/v2/oauth2'"
          }
        ]
      }
    },
    "/v2/oauth2?action=recoverauthorizationcode": {
      "post": {
        "tags": [
          "oauth2"
        ],
        "description": "This service allows you to retrieve a new authorization code in the event that the refresh token is invalid for a user belonging to your namespace. This service should not be used for standard end-user accounts. To generate a new authorization code for standard end-user accounts, please use the service [OAuth 2.0 - Get your authorization code](#operation/oauth2-authorize).\n\n**This service is limited to partners who are integrating Withings Mobile SDK or Withings Cellular solutions. This service will only work for partners having signed a contract with Withings.**",
        "summary": "OAuth 2.0 - Retrieve a new authorization code",
        "operationId": "oauth2-recoverauthorizationcode",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your ```Client ID```."
          },
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "A random token used to prevent replay attacks (Cf. [Signature v2 - Getnonce](#operation/signaturev2-getnonce))."
          },
          {
            "name": "signature",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hash of params (Cf. [Signature hash protocol](/developer-guide/v3/get-access/sign-your-requests))."
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User's email address"
          }
        ],
        "responses": {
          "200": {
            "description": " (Click to unfold)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "Response status. See <a href='#section/Response-status'>Status</a> section for details."
                    },
                    "body": {
                      "type": "object",
                      "description": "Response data.",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "description": "Authorization Code that must be used to retrieve access_token and refresh_token. (Read the \"Get access to your user's data\" section of your [integration guide](/developer-guide/v3/withings-solutions/integration-guides) to learn more about how to get this parameter.)",
                              "example": "490ed603fe9bd2ce10027bdba0c932069cd27085"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "PHP",
            "source": "<?php\n\n$ch = curl_init();\n\ncurl_setopt($ch, CURLOPT_URL, \"/v2/oauth2\");\n\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, TRUE);\n\ncurl_setopt($ch, CURLOPT_POSTFIELDS, http_build_query([ \n\t'action' => 'recoverauthorizationcode',\n\t'client_id' => '7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67',\n\t'signature' => '58c0c4ad271fb20df9b96c63dc68d6c550c0b7896f743262fc4c5b9261f32ab2',\n\t'nonce' => 'c455541b8f9111689adb647d32fac07a6ffe462d',\n\t'email' => 'recoverauthorizationcode@withings.com'\n]));\n\n$rsp = curl_exec($ch);\ncurl_close($ch);\n\nvar_dump($rsp);\n\n?>"
          },
          {
            "lang": "CURL",
            "source": "curl --data \"action=recoverauthorizationcode&client_id=7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67&signature=58c0c4ad271fb20df9b96c63dc68d6c550c0b7896f743262fc4c5b9261f32ab2&nonce=c455541b8f9111689adb647d32fac07a6ffe462d&email=recoverauthorizationcode@withings.com\" 'https://wbsapi.withings.net/v2/oauth2'"
          }
        ]
      }
    },
    "/v2/oauth2?action=getdemoaccess": {
      "post": {
        "tags": [
          "oauth2"
        ],
        "description": "This webservice gives your clientID access to a demo account containing dummy data",
        "summary": "OAuth 2.0 - Get access to demo account",
        "operationId": "oauth2-getdemoaccess",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Your ```Client ID```."
          },
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "A random token used to prevent replay attacks (Cf. [Signature v2 - Getnonce](#operation/signaturev2-getnonce))."
          },
          {
            "name": "scope_oauth2",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "A comma-separated list of permission scopes you want to ask your user for (see the [Index Data API](/developer-guide/v3/data-api/all-available-health-data) section to know which scope you should use)."
          }
        ],
        "responses": {
          "200": {
            "description": " (Click to unfold)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "description": "Response status. See <a href='#section/Response-status'>Status</a> section for details."
                    },
                    "body": {
                      "type": "object",
                      "description": "Response data.",
                      "properties": {
                        "access_token": {
                          "type": "string",
                          "description": "Your new Access Token.",
                          "example": "a075f8c14fb8df40b08ebc8508533dc332a6910a"
                        },
                        "refresh_token": {
                          "type": "string",
                          "description": "Your Refresh Token.",
                          "example": "f631236f02b991810feb774765b6ae8e6c6839ca"
                        },
                        "expires_in": {
                          "type": "integer",
                          "description": "Access token expiry delay in seconds.",
                          "example": 10800
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "PHP",
            "source": "<?php\n\n$ch = curl_init();\n\ncurl_setopt($ch, CURLOPT_URL, \"/v2/oauth2\");\n\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, TRUE);\n\ncurl_setopt($ch, CURLOPT_POSTFIELDS, http_build_query([ \n\t'action' => 'getdemoaccess',\n\t'client_id' => '7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67',\n\t'signature' => '58c0c4ad271fb20df9b96c63dc68d6c550c0b7896f743262fc4c5b9261f32ab2',\n\t'nonce' => 'c455541b8f9111689adb647d32fac07a6ffe462d',\n\t'scope_oauth2' => 'user.activity,user.metrics'\n]));\n\n$rsp = curl_exec($ch);\ncurl_close($ch);\n\nvar_dump($rsp);\n\n?>"
          },
          {
            "lang": "CURL",
            "source": "curl --data \"action=getdemoaccess&client_id=7573fd4a4c421ddd102dac406dc6e0e0e22f683c4a5e81ff0a5caf8b65abed67&signature=58c0c4ad271fb20df9b96c63dc68d6c550c0b7896f743262fc4c5b9261f32ab2&nonce=c455541b8f9111689adb647d32fac07a6ffe462d&scope_oauth2=user.activity,user.metrics\" 'https://wbsapi.withings.net/v2/oauth2'"
          }
        ]
      }
    },
...
  },
...
}